--- src/test/java/tz/go/moh/him/thscp/mediator/elmis/mock/MockDestination.java
@@ -1,59 +1,0 @@
-package tz.go.moh.him.thscp.mediator.elmis.mock;
-
-import org.apache.commons.io.IOUtils;
-import org.openhim.mediator.engine.messages.MediatorHTTPRequest;
-import org.openhim.mediator.engine.testing.MockHTTPConnector;
-import tz.go.moh.him.thscp.mediator.elmis.orchestrator.EmergencyCommodityStockStatusOrchestratorTest;
-
-import java.util.Collections;
-import java.util.Map;
-
-/**
- * Represents a mock destination.
- */
-public class MockDestination extends MockHTTPConnector {
-
-    /**
-     * Gets the response.
-     *
-     * @return Returns the response.
-     */
-    @Override
-    public String getResponse() {
-        try {
-            return IOUtils.toString(EmergencyCommodityStockStatusOrchestratorTest.class.getClassLoader().getResourceAsStream("success_response.json"));
-        } catch (Exception e) {
-            return null;
-        }
-    }
-
-    /**
-     * Gets the status code.
-     *
-     * @return Returns the status code.
-     */
-    @Override
-    public Integer getStatus() {
-        return 200;
-    }
-
-    /**
-     * Gets the HTTP headers.
-     *
-     * @return Returns the HTTP headers.
-     */
-    @Override
-    public Map<String, String> getHeaders() {
-        return Collections.emptyMap();
-    }
-
-    /**
-     * Handles the message.
-     *
-     * @param msg The message.
-     */
-    @Override
-    public void executeOnReceive(MediatorHTTPRequest msg) {
-        //TODO Implement asserts to check what has been received
-    }
-}
